+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 1.0.1
-- Fixed the settings of `codeQuicken.javascript.fileExtension`, `codeQuicken.javascript.indexFile`, and `codeQuicken.javascript.semiColons`.
-- Fixed the issue of `import * as * from "...";`.
+- Fixed the issue of `import * as * from "...";` in JavaScript.
+- Fixed the settings of `codeQuicken.javascript.fileExtension`, `codeQuicken.javascript.indexFile`, `codeQuicken.javascript.semiColons`, `codeQuicken.stylus.fileExtension`, `codeQuicken.stylus.indexFile`, and `codeQuicken.stylus.semiColons`.
 
 ### 1.0.0
 - Amended the whole extension to only support inserting and fixing `import`/`require` statements in JavaScript + TypeScript and Stylus langauges.